Add day 2 part 2 solution
Just doing the obvious but slow approach. I initially spent some time
trying to generalize the part 1 solution by summing the k-invalid
numbers for each k. Then I realized that counts invalid numbers multiple
times if k is not prime. Oops.

I think the approach could be salvaged using the principle of inclusion
exclusion (and would probably still be faster), but I decided to just do
this and call it a day.